% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Marker_visualizations.R
\name{extract_binned_expression}
\alias{extract_binned_expression}
\title{Title}
\usage{
extract_binned_expression(seurat_obj, gene_list, group_col = "seurat_clusters")
}
\arguments{
\item{group_col}{}
}
\description{
Title
}
