% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/General_visualizations.R
\name{plot_combined}
\alias{plot_combined}
\title{Title}
\usage{
plot_combined(
  seurat_obj,
  cluster_col,
  genes_list,
  sample_name = "Sample",
  reduction = "umap.cca",
  viridis_palette = "viridis",
  output_dir = "Plots_feat_dens",
  save_format = "jpg",
  width = 20,
  dpi = 300,
  add_timestamp = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{verbose}{}
}
\description{
Title
}
