% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/General_visualizations.R
\name{plot_gene_pair_correlations}
\alias{plot_gene_pair_correlations}
\title{Title}
\usage{
plot_gene_pair_correlations(
  seurat_obj,
  cluster_col,
  gene_pairs,
  sample_name = "Sample",
  cor_method = "pearson",
  quantile_thresh = 0.01,
  fill_palette = "auto",
  output_dir = "Plots_gene_pair_cor",
  save_format = "jpg",
  width = 14,
  height = 8,
  dpi = 300,
  add_timestamp = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{verbose}{}
}
\description{
Title
}
