% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Technical_noise_screening.R
\name{plot_vizdimloadings}
\alias{plot_vizdimloadings}
\title{Title}
\usage{
plot_vizdimloadings(
  seurat_obj,
  sample_name,
  reduction = "pca",
  pc_windows = list(1:10, 11:20, 21:30),
  output_dir = "VizDimLoadings_plots",
  nfeatures = 60,
  width_in = 20,
  height_in = 45,
  res = 300,
  ncol = 5
)
}
\arguments{
\item{ncol}{}
}
\description{
Title
}
