% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/General_visualizations.R
\name{PlotDimByFactors}
\alias{PlotDimByFactors}
\title{Title}
\usage{
PlotDimByFactors(
  seurat_obj,
  factors,
  sample_name = "Sample",
  reduction = "umap.har",
  raster = FALSE,
  label = TRUE,
  repel = TRUE,
  output_dir = "Plots_umap",
  plot_format = "jpg",
  width = 10,
  height = 10,
  dpi = 300,
  add_timestamp = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{verbose}{}
}
\description{
Title
}
