% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pre-processing_soupX.R
\name{prepare_soupx_inputs}
\alias{prepare_soupx_inputs}
\title{Title}
\usage{
prepare_soupx_inputs(filtered_dir, raw_dir, manual_contam = NULL)
}
\arguments{
\item{manual_contam}{}
}
\description{
Title
}
