% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pre-processing_soupX.R
\name{create_soup_channels}
\alias{create_soup_channels}
\title{Title}
\usage{
create_soup_channels(cts_raw_list, cts_filtered_list, sample_names)
}
\arguments{
\item{sample_names}{}
}
\description{
Title
}
