% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Marker_visualizations.R
\name{FindTopMarkersAndHeatmap}
\alias{FindTopMarkersAndHeatmap}
\title{Title}
\usage{
FindTopMarkersAndHeatmap(
  seurat_obj,
  sample_name = "Sample",
  marker_diff_thresh = 0.1,
  marker_pval_adj = 0.05,
  marker_avg_log2FC_thresh = 0.5,
  top_n = 20,
  output_dir_base = "Plots_heatmap",
  plot_format = "jpg",
  width = 10,
  height = 20,
  dpi = 300,
  compare = NULL,
  add_timestamp = TRUE,
  onlyPos = TRUE
)
}
\arguments{
\item{onlyPos}{}
}
\description{
Title
}
