% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSEA_ORA_DEseq2.R
\name{generate_and_save_heatmap}
\alias{generate_and_save_heatmap}
\title{Title}
\usage{
generate_and_save_heatmap(
  res_obj,
  comp_name,
  comp_title,
  vsd_data,
  anno_col,
  ordered_samps,
  n_padj,
  n_lfc,
  out_dir,
  ts,
  clus = "all"
)
}
\arguments{
\item{clus}{}
}
\description{
Title
}
