% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Technical_noise_screening.R
\name{generate_dimheatmaps}
\alias{generate_dimheatmaps}
\title{Title}
\usage{
generate_dimheatmaps(
  seurat_obj,
  sample_name,
  reduction = "pca",
  pc_windows = list(1:10, 11:20, 21:30),
  output_dir = "DimHeatmap_plots",
  nfeatures = 60,
  cells = 500,
  width_in = 15,
  height_in = 45,
  res = 300
)
}
\arguments{
\item{res}{}
}
\description{
Title
}
