% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stat_plot.R
\name{plot_cell_abundance}
\alias{plot_cell_abundance}
\title{Title}
\usage{
plot_cell_abundance(
  seurat_obj,
  sample_col,
  condition_col,
  celltype_col,
  global_test = "kruskal.test",
  strict_posthoc = TRUE,
  pairwise_test = "mann_whitney",
  p_adjust = "BH",
  facet_by_cluster = TRUE,
  facet_scales = "free_y",
  pairwise_label = "p.signif",
  y_expand = 0.2,
  output_dir = ".",
  base_size = 3.5,
  dpi = 300
)
}
\arguments{
\item{dpi}{}
}
\description{
Title
}
