% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Annotation_visualization.R
\name{plot_dot_dendro}
\alias{plot_dot_dendro}
\title{Title}
\usage{
plot_dot_dendro(
  seurat_obj,
  meta_col,
  feature_df,
  scale = TRUE,
  pct_threshold = 15,
  output_dir = "Output_R",
  prefix = NULL
)
}
\arguments{
\item{prefix}{}
}
\description{
Title
}
