% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Annotation_visualization.R
\name{plot_dot_dendro_split}
\alias{plot_dot_dendro_split}
\title{Title}
\usage{
plot_dot_dendro_split(
  seurat_obj,
  meta_col,
  feature_df,
  scale = TRUE,
  pct_threshold = 15,
  output_dir = "Output_R",
  prefix = NULL,
  max_genes_per_plot = NULL
)
}
\arguments{
\item{max_genes_per_plot}{}
}
\description{
Title
}
