% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Marker_visualizations.R
\name{plot_gene_markers_with_dendro}
\alias{plot_gene_markers_with_dendro}
\title{Title}
\usage{
plot_gene_markers_with_dendro(
  df,
  genes,
  id_col = "short_id",
  output_dir = "Markers_Plots",
  plot_title = "Marker Genes Plot",
  dendro_side = "right",
  label_space = 1,
  width_scale = 0.5,
  height_scale = 0.5,
  min_width = 5,
  min_height = 4
)
}
\arguments{
\item{min_height}{}
}
\description{
Title
}
