% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stat_plot.R
\name{plot_metadata_stats}
\alias{plot_metadata_stats}
\title{Title}
\usage{
plot_metadata_stats(
  seurat_obj,
  sample_col = "Sample",
  condition_col = "Severity",
  metadata_vars = c("Age", "Gender"),
  continuous_test_n2 = "mann_whitney",
  continuous_test_n3 = "kruskal.test",
  categorical_test = "chisq",
  strict_posthoc = TRUE,
  p_adjust = "BH",
  add_facet = NULL,
  output_dir = "metadata_plots",
  plot_width = 6,
  plot_height = 5,
  dpi = 300
)
}
\arguments{
\item{dpi}{}
}
\description{
Title
}
