% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Technical_noise_screening.R
\name{plot_pc_metadata_correlation}
\alias{plot_pc_metadata_correlation}
\title{Title}
\usage{
plot_pc_metadata_correlation(
  seurat_obj,
  sample_name,
  vars_to_test = c("pct_counts_mt", "nCount_RNA", "percent.ribo", "percent.ig",
    "nuclear_rna", "S.Score", "G2M.Score", "MALAT1", "NEAT1", "XIST"),
  reduction = "pca.log",
  n_pcs = 40,
  output_dir = "Output_R/Find_vartoregress",
  plot_width_in = 15,
  plot_height_in = 7,
  res_dpi = 300
)
}
\arguments{
\item{res_dpi}{}
}
\description{
Title
}
