% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Technical_noise_screening.R
\name{plot_technical_contribution}
\alias{plot_technical_contribution}
\title{Title}
\usage{
plot_technical_contribution(
  seurat_obj,
  sample_name,
  output_dir = "Output_R/Find_vartoregress",
  technical_keywords = c("^MT-", "^RPL", "^RPS", "^IG[HKL]", "MALAT1", "NEAT1", "XIST"),
  max_pcs = 40,
  n_top_genes = 500,
  cutoff = 15
)
}
\arguments{
\item{cutoff}{}
}
\description{
Title
}
