% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Processing.R
\name{ProcessSeuratLOG}
\alias{ProcessSeuratLOG}
\title{Title}
\usage{
ProcessSeuratLOG(
  seurat_obj,
  batch_col = "batch",
  vars_to_regress = NULL,
  tcr_bcr_patterns = "^TR[ABDG]|^IG[HKL]",
  reduction_name = "pca.SCT",
  integration_method = HarmonyIntegration,
  integration_reduction = "integrated.har.SCT",
  dims = 1:30,
  interactive_mode = FALSE,
  elbow_plot_dir = NULL,
  k.weight = NULL,
  k.anchor = NULL,
  k.filter = NULL,
  k.score = NULL,
  clustering_resolution = 1,
  verbose = TRUE,
  sample_name = "seurat"
)
}
\arguments{
\item{sample_name}{}
}
\description{
Title
}
