% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pre-processing_soupX.R
\name{run_soupx_post_clustering}
\alias{run_soupx_post_clustering}
\title{Title}
\usage{
run_soupx_post_clustering(
  prep_output,
  multiFac = 0,
  manual_contam = list(),
  pCut = 0.01,
  methods = "subtraction"
)
}
\arguments{
\item{methods}{}
}
\description{
Title
}
